requirements.txt
Flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
Werkzeug
tzdata
gunicorn